Move components into a new src folder
