+++ mobile-app/src/components/Board/Board.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import Svg from 'react-native-svg';
 import PlayableCell from '../PlayableCell/PlayableCell';
 import BorderCell from '../BorderCell/BorderCell';
-import { GameState, Coordinates } from "../../utils";
+import { GameState, Coordinates } from "../../../utils";
 import { getApproximateSvgSize, generateBoardCells, CELL_STROKE_COLOR } from "./BoardService";
 
 interface BoardProps {
