+++ mobile-app/src/components/BorderCell/BorderCell.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Polygon } from 'react-native-svg';
-import { RenderedCellType } from "../../utils";
+import { RenderedCellType } from "../../../utils";
 import { getBorderColor } from './BorderCellService';
 
 interface BorderCellProps {
